+++ src/com/topodroid/DistoX/GeoReference.java
@@ -18,10 +18,10 @@ class GeoReference
   final float e;
   final float s;
   final float v;
-  final float eradius;
-  final float sradius;
+  final double eradius;
+  final double sradius;
 
-  GeoReference( float e0, float s0, float v0, float er, float sr )
+  GeoReference( float e0, float s0, float v0, double er, double sr )
   {
     e = e0;
     s = s0;
