+++ src/com/topodroid/shp/ShpSegment.java
@@ -44,7 +44,7 @@ public class ShpSegment extends ShpObject
 
   // @param x0 x-offset
   // @param y0 y-offset
-  public boolean writeSegments( List< DrawingPath > sgms, float x0, float y0, float xscale, float yscale ) throws IOException
+  public boolean writeSegments( List< DrawingPath > sgms, double x0, double y0, double xscale, double yscale ) throws IOException
   {
     int nrs = ( sgms != null )? sgms.size() : 0;
     if ( nrs == 0 ) return false;
@@ -106,7 +106,7 @@ public class ShpSegment extends ShpObject
     return true;
   }
 
-  private void writeShpRecord( int cnt, int len, DrawingPath sgm, float x0, float y0, float xscale, float yscale )
+  private void writeShpRecord( int cnt, int len, DrawingPath sgm, double x0, double y0, double xscale, double yscale )
   {
     double xmin, ymin, xmax, ymax;
     {
@@ -142,7 +142,7 @@ public class ShpSegment extends ShpObject
   // @Override 
   protected int getShpRecordLength( ) { return 28 + 2 * 8; }
 
-  private void setBoundsLines( List< DrawingPath > sgms, float x0, float y0, float xscale, float yscale )
+  private void setBoundsLines( List< DrawingPath > sgms, double x0, double y0, double xscale, double yscale )
   {
     int nrs = ( sgms != null )? sgms.size() : 0;
     if ( nrs > 0 ) {
